feat: outsource for reuse
