+++ packages/app-config-writer/test/unit/common/utils.test.ts
@@ -1,0 +1,42 @@
+import { join } from 'path';
+import * as utils from '../../../src/common/utils';
+import { create as createFS } from 'mem-fs-editor';
+import { create as createStorage } from 'mem-fs';
+import type { Editor } from 'mem-fs-editor';
+
+describe('utils', () => {
+    const basePath = join(__dirname, '../../fixtures/variants-config');
+
+    describe('getPreviewMiddleware', () => {
+        test('exception handling - parameters not provided', async () => {
+            await expect(utils.getPreviewMiddleware()).rejects.toThrowError(
+                'Either base path or yaml config must be provided'
+            );
+        });
+
+        test('exception handling - file not found', async () => {
+            const basePath = join(__dirname, '../../fixtures/a-folder-that-does-not-exist');
+            await expect(utils.getPreviewMiddleware(undefined, basePath, 'chicken.html')).rejects.toThrowError(
+                `File 'chicken.html' not found in project '${basePath}'`
+            );
+        });
+    });
+
+    describe('getRTAServe', () => {
+        let fs: Editor;
+
+        beforeEach(() => {
+            jest.clearAllMocks();
+            fs = createFS(createStorage());
+        });
+        test('RTA serve for preview middleware', async () => {
+            const openSourceConfig = join(basePath, 'open-source-config');
+            expect(await utils.getRTAServe(openSourceConfig, 'ui5.yaml', fs)).toStrictEqual('ui5 serve');
+        });
+
+        test('RTA serve for fiori-tools-preview middleware', async () => {
+            const fioriToolsConfig = join(basePath, 'fiori-tools-config');
+            expect(await utils.getRTAServe(fioriToolsConfig, 'ui5.yaml', fs)).toStrictEqual('fiori run');
+        });
+    });
+});
